+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/SdkSynchronizer.kt
@@ -32,6 +32,7 @@ import cash.z.ecc.android.sdk.db.entity.isSubmitted
 import cash.z.ecc.android.sdk.exception.SynchronizerException
 import cash.z.ecc.android.sdk.ext.ConsensusBranchId
 import cash.z.ecc.android.sdk.ext.ZcashSdk
+import cash.z.ecc.android.sdk.internal.SaplingParamTool
 import cash.z.ecc.android.sdk.internal.block.CompactBlockDbStore
 import cash.z.ecc.android.sdk.internal.block.CompactBlockDownloader
 import cash.z.ecc.android.sdk.internal.block.CompactBlockStore
@@ -73,11 +74,9 @@ import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.cancel
-import kotlinx.coroutines.channels.ConflatedBroadcastChannel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.first
@@ -103,7 +102,6 @@ import kotlin.coroutines.EmptyCoroutineContext
  * @property processor saves the downloaded compact blocks to the cache and then scans those blocks for
  * data related to this wallet.
  */
-@OptIn(kotlinx.coroutines.ObsoleteCoroutinesApi::class)
 @FlowPreview
 @Suppress("TooManyFunctions")
 class SdkSynchronizer internal constructor(
@@ -117,9 +115,7 @@ class SdkSynchronizer internal constructor(
     private val _saplingBalances = MutableStateFlow<WalletBalance?>(null)
     private val _transparentBalances = MutableStateFlow<WalletBalance?>(null)
 
-    // TODO [#288]: Remove Deprecated Usage of ConflatedBroadcastChannel
-    // TODO [#288]: https://github.com/zcash/zcash-android-wallet-sdk/issues/288
-    private val _status = ConflatedBroadcastChannel<Synchronizer.Status>(DISCONNECTED)
+    private val _status = MutableStateFlow<Synchronizer.Status>(DISCONNECTED)
 
     /**
      * The lifespan of this Synchronizer. This scope is initialized once the Synchronizer starts
@@ -181,10 +177,7 @@ class SdkSynchronizer internal constructor(
      * processor is finished scanning, the synchronizer updates transaction and balance info and
      * then emits a [SYNCED] status.
      */
-    // TODO [#658] Replace ComputableFlow and asFlow() obsolete Coroutine usage
-    // TODO [#658] https://github.com/zcash/zcash-android-wallet-sdk/issues/658
-    @Suppress("DEPRECATION")
-    override val status = _status.asFlow()
+    override val status = _status.asStateFlow()
 
     /**
      * Indicates the download progress of the Synchronizer. When progress reaches 100, that
@@ -310,7 +303,7 @@ class SdkSynchronizer internal constructor(
             twig("Synchronizer::stop: coroutineScope.cancel()")
             coroutineScope.cancel()
             twig("Synchronizer::stop: _status.cancel()")
-            _status.cancel()
+            _status.value = STOPPED
             twig("Synchronizer::stop: COMPLETE")
         }
     }
@@ -420,7 +413,7 @@ class SdkSynchronizer internal constructor(
                 //  ignore enhancing status for now
                 // TODO [#682]: clean this up and handle enhancing gracefully
                 // TODO [#682]: https://github.com/zcash/zcash-android-wallet-sdk/issues/682
-                if (synchronizerStatus != ENHANCING) _status.send(synchronizerStatus)
+                if (synchronizerStatus != ENHANCING) _status.value = synchronizerStatus
             }
         }.launchIn(this)
         processor.start()
@@ -817,9 +810,9 @@ object DefaultSynchronizerFactory {
         val coordinator = DatabaseCoordinator.getInstance(context)
 
         return RustBackend.init(
-            coordinator.cacheDbFile(network, alias).absolutePath,
-            coordinator.dataDbFile(network, alias).absolutePath,
-            File(context.getCacheDirSuspend(), "params").absolutePath,
+            coordinator.cacheDbFile(network, alias),
+            coordinator.dataDbFile(network, alias),
+            File(context.getCacheDirSuspend(), "params"),
             network,
             blockHeight
         )
@@ -864,8 +857,9 @@ object DefaultSynchronizerFactory {
 
     internal fun defaultEncoder(
         rustBackend: RustBackend,
+        saplingParamTool: SaplingParamTool,
         repository: TransactionRepository
-    ): TransactionEncoder = WalletTransactionEncoder(rustBackend, repository)
+    ): TransactionEncoder = WalletTransactionEncoder(rustBackend, saplingParamTool, repository)
 
     fun defaultDownloader(
         service: LightWalletService,