+++ CHANGELOG.md
@@ -71,6 +71,17 @@ Change Log
   - `DerivationTool.deriveUnifiedViewingKeys`
   - `DerivationTool.validateUnifiedViewingKey`
 
+Version 1.9.0-beta05
+------------------------------------
+- The minimum version of Android supported is now API 21
+- Fixed R8/ProGuard consumer rule, which eliminates a runtime crash for minified apps
+
+Version 1.9.0-beta04
+------------------------------------
+- The SDK now stores sapling param files in `no_backup/co.electricoin.zcash` folder instead of the `cache/params` 
+  folder. Besides that, `SaplingParamTool` also does validation of downloaded sapling param file hash and size.
+**No action required from client app**.
+
 Version 1.9.0-beta03
 ------------------------------------
 - No changes; this release is a test of a new deployment process